requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-login==0.6.3
Flask-SQLAlchemy==3.1.1
flask-wtf==1.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
wtforms==3.1.2
SQLAlchemy==2.0.28
